requirements.txt

fastapi==0.110.0
uvicorn[standard]
sqlalchemy==1.4.54
psycopg2-binary
python-dotenv
pydantic<2.0
passlib[bcrypt]
python-jose[cryptography]
email-validator
alembic
